+++ MegaManX3/MegaManX3/GameSence.h
@@ -6,8 +6,8 @@ class GameSence
 {
 public:
 	virtual void LoadContent() = 0;
-	virtual void Update(float time) = 0;
-	virtual void Draw(float time) = 0;
+	virtual void Update(double time) = 0;
+	virtual void Draw(double time) = 0;
 };
 
 #endif //__GAMESENCE__H__
